move code to redacted.rs, move tests and remove dbg statements
