+++ icann-rdap-client/Cargo.toml
@@ -27,6 +27,9 @@ strum.workspace = true
 strum_macros.workspace = true
 thiserror.workspace = true
 
+jsonpath-rust = "=0.5.0"
+jsonpath_lib = "0.3.0"
+
 [dev-dependencies]
 
 # fixture testings
