+++ icann-rdap-client/src/test_files/example-2_partial_value.json
