+++ icann-rdap-common/Cargo.toml
@@ -24,10 +24,6 @@ strum.workspace = true
 strum_macros.workspace = true
 thiserror.workspace = true
 
-jsonpath-rust = "=0.5.0"
-jsonpath_lib = "0.3.0"
-regex = "1.5.4"
-
 [dev-dependencies]
 
 # fixture testings
